feat: display query parameters in the browser url
