+++ react-project/src/App.scss
@@ -9,22 +9,6 @@
   height: 100vh;
 }
 
-.app {
-  position: relative;
-  margin-top: 30px;
-
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-
-  .error {
-    position: fixed;
-    z-index: 2;
-    bottom: 50px;
-    right: 100px;
-  }
-}
-
 button {
   width: 200px;
   height: 60px;
