+++ react-project/src/pages/Main/MainPage.scss
@@ -1,0 +1,15 @@
+.main-page {
+  position: relative;
+  margin-top: 30px;
+
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+
+  .error {
+    position: fixed;
+    z-index: 2;
+    bottom: 50px;
+    right: 100px;
+  }
+}
