+++ react-project/src/pages/NotFound/NotFoundPage.tsx
@@ -1,0 +1,5 @@
+function NotFound() {
+  return <h1>Not found</h1>;
+}
+
+export default NotFound;
